deprecated sqlwarehouse_id param
